+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - '-p' flag to prompt on all task changing commands
+- Added "addx" command to add completed tasks
 
 ## [0.4.0] - 2018-04-18
 ### Changed